+++ packages/melody-hooks/__tests__/util/flush.js
@@ -1,0 +1,8 @@
+import { flush as _flush } from 'melody-idom';
+export const flush = () =>
+    _flush({
+        didTimeout: false,
+        timeRemaining() {
+            return 10;
+        },
+    });
